fix: 컨벤션에 맞게 수정 및 불필요한 코드 삭제
